Buffer test

+++ src/filter/dbfilter.cpp
@@ -302,10 +302,10 @@ bool DBFilter::activeInLiveMode()
     return false;
 }
 
-std::vector<size_t> DBFilter::filterBuffer(const std::string& dbcontent_name, std::shared_ptr<Buffer> buffer)
+std::vector<unsigned int> DBFilter::filterBuffer(const std::string& dbcontent_name, std::shared_ptr<Buffer> buffer)
 {
     assert (activeInLiveMode()); // re-implement in sub-class
-    return std::vector<size_t>();
+    return std::vector<unsigned int>();
 }
 
 bool DBFilter::widgetVisible() const
