+++ src/filter/modecfilter.cpp
@@ -194,9 +194,9 @@ bool ModeCFilter::activeInLiveMode()
     return true;
 }
 
-std::vector<size_t> ModeCFilter::filterBuffer(const std::string& dbcontent_name, std::shared_ptr<Buffer> buffer)
+std::vector<unsigned int> ModeCFilter::filterBuffer(const std::string& dbcontent_name, std::shared_ptr<Buffer> buffer)
 {
-    std::vector<size_t> to_be_removed;
+    std::vector<unsigned int> to_be_removed;
 
     if (!COMPASS::instance().dbContentManager().metaVariable(DBContent::meta_var_mc_.name()).existsIn(dbcontent_name))
         return to_be_removed;
